refactor: use gin in crash_reporter
