+++ shell/app/node_main.cc
@@ -16,13 +16,12 @@
 #include "gin/array_buffer.h"
 #include "gin/public/isolate_holder.h"
 #include "gin/v8_initializer.h"
-#include "native_mate/dictionary.h"
 #include "shell/app/uv_task_runner.h"
 #include "shell/browser/javascript_environment.h"
 #include "shell/browser/node_debugger.h"
 #include "shell/common/api/electron_bindings.h"
 #include "shell/common/crash_reporter/crash_reporter.h"
-#include "shell/common/native_mate_converters/string16_converter.h"
+#include "shell/common/gin_helper/dictionary.h"
 #include "shell/common/node_bindings.h"
 #include "shell/common/node_includes.h"
 
@@ -94,14 +93,15 @@ int NodeMain(int argc, char* argv[]) {
 
     node::BootstrapEnvironment(env);
 
-    mate::Dictionary process(gin_env.isolate(), env->process_object());
+    gin_helper::Dictionary process(gin_env.isolate(), env->process_object());
 #if defined(OS_WIN)
     process.SetMethod("log", &ElectronBindings::Log);
 #endif
     process.SetMethod("crash", &ElectronBindings::Crash);
 
     // Setup process.crashReporter.start in child node processes
-    auto reporter = mate::Dictionary::CreateEmpty(gin_env.isolate());
+    gin_helper::Dictionary reporter =
+        gin::Dictionary::CreateEmpty(gin_env.isolate());
     reporter.SetMethod("start", &crash_reporter::CrashReporter::StartInstance);
 
 #if !defined(OS_LINUX)
@@ -111,7 +111,7 @@ int NodeMain(int argc, char* argv[]) {
 
     process.Set("crashReporter", reporter);
 
-    mate::Dictionary versions;
+    gin_helper::Dictionary versions;
     if (process.Get("versions", &versions)) {
       versions.SetReadOnly(ELECTRON_PROJECT_NAME, ELECTRON_VERSION_STRING);
     }
